+++ src/App.js
@@ -1,7 +1,8 @@
 import React, {useState,useEffect,useRef} from 'react';
 import Ticker from './components/Ticker';
 import Display from './components/Display';
-import {useIsFirstRender} from './hooks/useIsFirstRender';
+import useIsFirstRender from './hooks/useIsFirstRender';
+
 const App = () => {
     //session/break lengths (in minutes)
     const [seshVal,setSeshVal]=useState(25);
@@ -13,45 +14,94 @@ const App = () => {
     //current timer value (in s)
     const [curr,setCurr]=useState(1500);
 
-    //interval reference
-    const [basis,setBasis]=useState(-1);
-    const [intRef,setIntRef]=useState(null);
-
     //audio reference
     const audio = useRef(null);
 
     //first render state
     const isFR=useIsFirstRender();
+
+    //for timer
+    //expected time lapse: the expected date.now() value when next settimeout is executed
+    //the timeout value for next settimeout
+    const [timeItems,updateTimeItems]=useState({
+        expTimeLapse:0,
+        nextUp:1000
+    });
+    //timeout reference for stopping
+    const [timeRef,setTimeRef]=useState(0);
+
     //object storing all functions related to changing timer lengths
     const timeChange={
         //increment session length by 1 if is paused and result would not go over 60
         handleSInc:()=>{        
-            if (basis===-1 && seshVal+1<=60) 
+            if (timeItems.expTimeLapse===0 && seshVal+1<=60) 
                 setSeshVal(seshVal+1);
         },
         //decrement session length by 1 if is paused and result would not go under 0
         handleSDec:()=>{        
-            if (basis===-1 && seshVal-1>0) 
+            if (timeItems.expTimeLapse===0 && seshVal-1>0) 
                 setSeshVal(seshVal-1);
         },
         //increment break length by 1 if is paused and result would not go over 60
         handleBInc:()=>{        
-            if (basis===-1 && breakVal+1<=60)
+            if (timeItems.expTimeLapse===0 && breakVal+1<=60)
                 setBreakVal(breakVal+1);
         },
         //decrement break length by 1 if is paused and result would not go under 0
         handleBDec:()=>{        
-            if (basis===-1 && breakVal-1>0)
+            if (timeItems.expTimeLapse===0 && breakVal-1>0)
                 setBreakVal(breakVal-1);
         }
     }
 
-    //changes basis state
+    //decrement current function used as callback for check
+    const updateCurr=()=>setCurr(curr-1);
+
+    //when timer running, call every settimeout
+    const check=()=>{
+        //calculating drift from expected time lapsed
+        const d=Date.now()-timeItems.expTimeLapse;
+        //if drift is larger than 1 second, 
+        //consider: immediately run next settimeout to catch up
+        //or: stop timer?
+        if (d>1000) {
+            console.log(d,1000,'skipped time!');
+        }
+        updateCurr();
+        //current solution: immediately run next settimeout to catch up
+        //update the list of timer items to new ones
+        updateTimeItems({
+            expTimeLapse:timeItems.expTimeLapse+1000,
+            nextUp:1000>=d?1000-d:0
+        });
+    }
+
+    //runs when no timer is running 
+    //settimeout handled by useeffect
+    const startTimer=()=>{
+        updateTimeItems({
+            expTimeLapse:Date.now()+1000,
+            nextUp:1000
+        });
+    }
+
+    //stops timer when one is running
+    //cleartimeout handled by useeffect
+    const stopTimer=()=>{
+        updateTimeItems({
+            expTimeLapse:0,
+            nextUp:1000
+        });
+    }
+
+    //checks if to start or pause timer by checking expectd time lapsed value
     const handleTimerAction=()=>{
-        if (basis===-1) 
-            setBasis(new Date());
-        else 
-            setBasis(-1);
+        if (timeItems.expTimeLapse===0) {
+            startTimer();
+        }
+        else {
+            stopTimer();
+        }
     };
 
     //handles complete reset
@@ -61,8 +111,8 @@ const App = () => {
         setBreakVal(5);
         //useeffect would be unreliable since timer on default setting being reset would not trigger it
         setCurr(1500);
-        //useeffect would clearinterval here
-        setBasis(-1);
+        //cleartimeout handled by useeffect
+        stopTimer();
         //stop alarm if playing
         audio.current.pause();
         audio.current.currentTime=0;
@@ -76,55 +126,55 @@ const App = () => {
         return (minute<10?`0${minute}`:minute)+':'+(second<10?`0${second}`:second);
     };
 
-
-    //counts down timer value
-    const calc=(rem,base)=>setCurr(rem-~~((Date.now()-base)/1000));
-
+    //when timer items changes:
+    //if expected time lapsed is 0, means timer is now paused, respond by clear running timer and resetting the timeout reference
+    //else the next settimeout should be set to check against new expected time lapsed, respond by settimeout
     useEffect(()=>{
         if (isFR)
             return;
-        
-            if (basis===-1) {
-                clearInterval(intRef);
-                setIntRef(null);
-            }
-            else {
-                setIntRef(setInterval(()=>calc(curr,basis),1000));
-            }
-    },[basis]);
+        if (timeItems.expTimeLapse===0) {
+            clearTimeout(timeRef);
+            setTimeRef(0);
+        }
+        else {
+            setTimeRef(setTimeout(check,timeItems.nextUp));
+        }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    },[timeItems]);
+
     //when session value changes, check if necessary to change curr timer value
     useEffect(()=>{
         if (isFR)
             return;
         console.log('seshval.breakval,issesh state change triggered useeffect');
-        console.log(`issesh ${isSesh}, seshVal ${seshVal}, breakVal ${breakVal}`);
+        console.log(`issesh ${isSesh}, seshVal ${seshVal}, breakVal ${breakVal}, curr ${curr}`);
         if (isSesh)
             setCurr(seshVal*60);
         else
             setCurr(breakVal*60);
-    },[isSesh,seshVal,breakVal]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    },[seshVal,breakVal]);
 
-            
-    const handleSwap=currTimerState=>{
-        setIsSesh(!currTimerState);
-    }
-    //when curr value changes, check if necessary to swap session/break
+    //in charge of what to do when curr passes certain threshold
     useEffect(()=>{
-        if (isFR)
-            return;
-
-            console.log('curr state change triggered useeffect');
-            console.log(`curr ${curr}`);
+        //5 seconds left, play blinking animation
         if (curr===5) {
-            //play animation
+            console.log('animation');
         }
+        //timers up, play alarm
         else if (curr===0) {
             audio.current.currentTime=0;
             audio.current.play();
         }
+        //last second up, swap timer phase
         else if (curr<0) {
-            handleSwap(isSesh);
+            const nextState=!isSesh;
+            stopTimer();
+            setIsSesh(nextState);
+            setCurr(nextState?seshVal*60:breakVal*60);
+            startTimer();
         }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[curr]);
 
     return (
@@ -140,7 +190,7 @@ const App = () => {
                 handleInc={timeChange.handleBInc}
                 handleDec={timeChange.handleBDec}/>
             <Display isSesh={isSesh}
-                isPause={basis===-1}
+                isPause={timeItems.expTimeLapse===0}
                 value={clock(curr)}
                 handleTimerAction={handleTimerAction}
                 handleReset={handleReset} />