+++ src/hooks/useIsFirstRender.js
@@ -1,9 +1,11 @@
 import { useRef, useEffect } from 'react';
 
-export const useIsFirstRender = () => {
+const useIsFirstRender = () => {
   const r = useRef(true);
   useEffect(() => {
     r.current = false;
   }, []);
   return r.current;
 };
+
+export default useIsFirstRender;